requirements.txt
aiofiles==23.2.1
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alpha_vantage==3.0.0
altair==5.5.0
annotated-types==0.7.0
anyio==3.7.1
asyncer==0.0.7
attrs==24.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
cachetools==5.5.1
certifi==2024.12.14
chainlit==2.0.5
charset-normalizer==3.4.1
chevron==0.14.0
click==8.1.8
contourpy==1.3.1
cycler==0.12.1
dataclasses-json==0.6.7
dateparser==1.2.0
Deprecated==1.2.15
distro==1.9.0
fastapi==0.115.7
fastapi-socketio==0.0.10
filetype==1.2.0
fonttools==4.55.4
frozendict==2.4.6
frozenlist==1.5.0
gitdb==4.0.12
GitPython==3.1.44
googleapis-common-protos==1.66.0
GoogleNews==1.6.15
grpcio==1.69.0
h11==0.14.0
html5lib==1.1
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
idna==3.10
importlib_metadata==8.5.0
Jinja2==3.1.5
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kiwisolver==1.4.8
langchain==0.3.15
langchain-community==0.3.15
langchain-core==0.3.31
langchain-openai==0.3.1
langchain-text-splitters==0.3.5
langsmith==0.2.11
Lazify==0.4.0
literalai==0.1.103
lxml==5.3.0
markdown-it-py==3.0.0
MarkupSafe==3.0.2
marshmallow==3.25.1
matplotlib==3.10.0
mdurl==0.1.2
multidict==6.1.0
multitasking==0.0.11
mypy-extensions==1.0.0
narwhals==1.23.0
nest-asyncio==1.6.0
numpy==1.26.4
openai==1.59.9
opentelemetry-api==1.28.2
opentelemetry-exporter-otlp==1.28.2
opentelemetry-exporter-otlp-proto-common==1.28.2
opentelemetry-exporter-otlp-proto-grpc==1.28.2
opentelemetry-exporter-otlp-proto-http==1.28.2
opentelemetry-instrumentation==0.49b2
opentelemetry-proto==1.28.2
opentelemetry-sdk==1.28.2
opentelemetry-semantic-conventions==0.49b2
orjson==3.10.15
packaging==23.2
pandas==2.2.3
peewee==3.17.8
pillow==11.1.0
platformdirs==4.3.6
propcache==0.2.1
protobuf==5.29.3
pyarrow==19.0.0
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.11.2
python-multipart==0.0.18
python-socketio==5.12.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.36.1
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==13.9.4
rpds-py==0.22.3
simple-websocket==1.1.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.37
starlette==0.41.3
streamlit==1.41.1
syncer==2.0.3
tenacity==9.0.0
tiktoken==0.8.0
toml==0.10.2
tomli==2.2.1
tornado==6.4.2
tqdm==4.67.1
typing==3.7.4.3
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uptrace==1.28.2
urllib3==2.3.0
uvicorn==0.25.0
watchfiles==0.20.0
webencodings==0.5.1
wikipedia==1.4.0
wrapt==1.17.2
wsproto==1.2.0
yarl==1.18.3
yfinance==0.2.52
zipp==3.21.0
